Separate api state
