--- frontend/src/business/models/Api.ts
@@ -1,5 +1,0 @@
-export interface ApiResponse {
-  statusText?: string;
-  message: string;
-  status: "success" | "error" | undefined;
-}
