+++ frontend/src/business/overmind/Api/state.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-shadow */
+import { ApiState } from "../../models/State";
+
+export const state: ApiState = {
+  response: { statusText: "", message: "", status: undefined },
+};
