+++ frontend/src/business/overmind/Review/actions.ts
@@ -6,11 +6,11 @@ export const getReviews = async (
   { actions, effects, state }: Context,
   { limit, page }: { limit: number; page: number }
 ) => {
-  actions.review.resetApiResponse();
+  actions.api.resetApiResponse();
   const response = await effects.review.api.getReviews({ limit, page });
 
   if (!response || response instanceof AxiosError) {
-    actions.review.handleErrorResponse({ response });
+    actions.api.handleErrorResponse({ response });
     return;
   }
 
@@ -40,46 +40,15 @@ export const postReview = async (
     token
   );
   if (response.status === "success") {
-    state.apiResponse = {
+    state.api.response = {
       message: response.message,
       status: "success",
     };
   } else {
-    state.apiResponse = {
+    state.api.response = {
       statusText: "Bad request",
       message: response.message,
       status: "error",
     };
   }
-};
-
-export const handleErrorResponse = (
-  { state }: Context,
-  { response }: { response: unknown }
-) => {
-  if (
-    response &&
-    response instanceof AxiosError &&
-    response.response &&
-    "data" in response.response &&
-    typeof response.response.data === "object" &&
-    response.response.data != null &&
-    "detail" in response.response.data
-  ) {
-    state.apiResponse = {
-      statusText: JSON.stringify(response.response.statusText),
-      message: JSON.stringify(response.response?.data.detail),
-      status: "error",
-    };
-  } else {
-    state.apiResponse = {
-      statusText: "Bad request",
-      message: "something went very wrong",
-      status: "error",
-    };
-  }
-};
-
-export const resetApiResponse = ({ state }: Context) => {
-  state.apiResponse = { statusText: "", message: "", status: undefined };
 };
