+++ frontend/src/business/overmind/index.ts
@@ -6,6 +6,7 @@ import {
   createStateHook,
 } from "overmind-react";
 import { namespaced } from "overmind/config";
+import * as api from "./Api";
 import * as auth from "./Auth";
 import * as author from "./Author";
 import * as book from "./Book";
@@ -17,6 +18,7 @@ import * as review from "./Review";
 // import { state } from "./state";
 
 export const config = namespaced({
+  api,
   auth,
   author,
   book,
